// src/types/bindings/CustomerMasterRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CustomerMasterRow = { id: bigint, customer_code: string, report_name: string, tally_name: string | null, legal_name: string | null, gstin: string | null, address1: string | null, address2: string | null, location: string | null, pincode: string | null, state_code: string | null, place_of_supply: string | null, phone: string | null, email: string | null, category_name: string | null, remarks: string | null, status: string, match_status: string, };